Add files via upload
Alright so things were acting up because of how you uploaded everything Zach so I went ahead and re-organized the file so everything is in the SRC folder.  Infix calc is not activating for some reason but that seems to be the only error If we can configure the program we will be good.